seed(issue-3): mean truncation + whitespace not stripped

+++ src/demo_app/calculator.py
@@ -26,7 +26,8 @@ def mean(values: list[float]) -> float:
     """
     if not values:
         raise ValueError("mean() requires at least one value")
-    return sum(values) / len(values)
+    # BUG (seed/issue-3): mistakenly truncates to int.
+    return float(int(sum(values) / len(values)))
 
 
 @dataclass(frozen=True)
